pyproject.toml
[project]
name = "api-router"
version = "0.1.0"
description = "Lightweight API proxy with multi-tenant auth and upstream routing"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Student team"}
]

dependencies = [
    "fastapi[standard]>=0.115.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "agent-robert",
]

[tool.uv.sources]
agent-robert = { path = "../agent-robert/agent", editable = true }


[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api_router"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
